Disable compilation of crates that use OptiX.
Because they don't work without OptiX, which is less important than CUDA
and a pain to install.

This means we don't have to manually exclude all these crates in CI,
which is nice -- e.g. a vanilla `cargo build --workspace` works.

Also, on Linux we don't need `--exclude "cudnn*"`. (We still need it on
Windows.)