Remove `shared_array!`
It's wildly unsound, as far as I can tell.
- The use of `UnsafeCell`/`unsafe impl Send`/`unsafe impl Sync` is very
  dodgy.
- The use of `MaybeUninit<[T; N]>` is completely wrong. It should be
  `[<MaybeUninit<T>; N>]`, which allows for partial initialization.
- The examples don't use `write`/`assume_init` with `MaybeUninit`, which
  is UB.

The alternative is to use `#[address_space(shared) static mut
[MaybeUninit<T>; N]` directly instead. That avoids all the unsoundness,
and is more ergonomic because you don't have to work with a raw pointer,
and is clearer because details aren't hidden within the macro.

I moved (with some modifications) the comments on `shared_array` to the
`address_space` proc macro because there were some useful details in
there.