cuda_std: use `usize` in `BlockSize` and `GridSize`.
These types are very similar to the samed-named types from `cust`, which
were changed in the previous commit. Currently these types are in a
module that is commented out and marked as "WIP", but it makes sense to
change them like the `cust` types in case they become used in the
future. (Note: I temporarily uncommented the code to make sure the
changes compile.)